requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
datasets>=2.14.0
trl>=0.7.0

# Astronomical Calculations
pyswisseph>=2.10.3.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
PyPDF2>=3.0.0
pymupdf>=1.23.0  # Better PDF extraction (PyMuPDF/fitz)

# API and Server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0
requests>=2.31.0

# Logging and Monitoring
tensorboard>=2.14.0
wandb>=0.16.0

# Development Tools
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0

# Optional: For distributed training
deepspeed>=0.12.0